Move errorProcessing util into the formatRPCResponse
